lectures/day_9/requirements.txt
pandas==2.1.0
apache-airflow-providers-postgres
psycopg2-binary